# cmd/pokeui/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PokePilot operator HTTP",
    "version": "1",
    "description": "Allowlisted operator REST on this origin. MCP is a separate Streamable HTTP server at /mcp and requires Authorization: Bearer POKEPILOT_MCP_TOKEN; it cannot proxy these paths. Operator REST does not use that bearer. Run ids include the run- prefix."
  },
  "paths": {
    "/v1/runs/{id}": {
      "get": {
        "summary": "Inspect one run",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Run plus compact finish metadata" },
          "404": { "description": "Unknown run" }
        }
      }
    },
...
  }
}